Extract commands to enum
