+++ chat-widget/src/components/webchatcontainerstateful/common/MockBotCardCommandType.ts
@@ -1,0 +1,7 @@
+enum MockBotCardCommandType {
+    Hero = "hero",
+    Signin = "signin",
+    Thumbnail = "thumbnail"
+}
+
+export default MockBotCardCommandType;
