+++ chat-widget/src/components/webchatcontainerstateful/common/MockBotCommand.ts
@@ -1,0 +1,12 @@
+enum MockBotCommand {
+    Bot = "/bot",
+    Card = "/card",
+    Help = "/help",
+    SendAttachment = "send attachment",
+    SendBotMessage = "send bot message",
+    SendSystemMessage = "send system message",
+    SendTyping = "send typing",
+    System = "/system"
+}
+
+export default MockBotCommand;
